+++ server/src/main/java/org/opensearch/extensions/ExtensionsSettings.java
@@ -8,8 +8,6 @@
 
 package org.opensearch.extensions;
 
-import org.opensearch.common.settings.Settings;
-
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;